새 폴더/10-01-django-rest-framework/assignment/django_ws_10_3/requirements.txt
asgiref==3.7.2
Django==4.2.11
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
